requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
beautifulsoup4==4.8.0
certifi==2019.6.16
chardet==3.0.4
cycler==0.10.0
discord.py==1.2.3
idna==2.8
kiwisolver==1.1.0
matplotlib==3.1.1
multidict==4.5.2
numpy==1.17.1
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.2
requests==2.22.0
schedule==0.6.0
six==1.12.0
soupsieve==1.9.3
urllib3==1.25.3
websockets==6.0
yarl==1.3.0
